requirements.txt
# Install with: pip install -r requirements.txt
# Or: pip install -e .  (editable install from pyproject.toml)
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0
pydantic-settings>=2.0
httpx>=0.27.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
chromadb>=0.5.0
python-dotenv>=1.0.0
